requirements.txt
discord.py==1.4.0
asyncpg==0.21.0
flake8==3.8.3
flake8-quotes==3.2.0
